# Move Crockford encoding and decoding chars to functions for performance