Issue/Task Number:

# Overview

The encoding and decoding in the Crockford module is sped up by using functions for the chars in the encoding scheme. This is similar to the Unicode implementation used in the Elixir standard library.

# Benchmark
## Baseline
```
Operating System: Linux
CPU Information: AMD Ryzen 7 4800H with Radeon Graphics
Number of Available Cores: 16
Available memory: 15.05 GB
Elixir 1.16.3
Erlang 25.3.2.12
Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
parallel: 1
inputs: none specified
Estimated total run time: 7 s


Benchmarking encode...
Type help for instructions on how to use fish
Name             ips        average  deviation         median         99th %
encode      124.07 K        8.06 μs   ±530.89%           7 μs          16 μs
Operating System: Linux
CPU Information: AMD Ryzen 7 4800H with Radeon Graphics
Number of Available Cores: 16
Available memory: 15.05 GB
Elixir 1.16.3
Erlang 25.3.2.12
Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
parallel: 1
inputs: none specified
Estimated total run time: 7 s


Benchmarking decode...

Name             ips        average  deviation         median         99th %
decode       47.57 K       21.02 μs    ±36.88%          21 μs          47 μs
```
## Improved
```
Operating System: Linux
CPU Information: AMD Ryzen 7 4800H with Radeon Graphics
Number of Available Cores: 16
Available memory: 15.05 GB
Elixir 1.16.3
Erlang 25.3.2.12
Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
parallel: 1
inputs: none specified
Estimated total run time: 7 s


Benchmarking encode...

Name             ips        average  deviation         median         99th %
encode      147.27 K        6.79 μs   ±633.59%           6 μs          13 μs
Operating System: Linux
CPU Information: AMD Ryzen 7 4800H with Radeon Graphics
Number of Available Cores: 16
Available memory: 15.05 GB
Elixir 1.16.3
Erlang 25.3.2.12
Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
parallel: 1
inputs: none specified
Estimated total run time: 7 s


Benchmarking decode...

Name             ips        average  deviation         median         99th %
decode      363.11 K        2.75 μs  ±2540.59%           2 μs           6 μs
```